upload-broadcast-service/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
pysolr~=3.9.0